requirements.txt
click==7.1.2
Flask==1.1.2
importlib-metadata==4.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
python-dotenv==0.19.0
Werkzeug==1.0.1
zipp==3.5.0
